Removed libpng (and hopefully zlib) dependency
Fixed segfault caused by flipped conditional in save_png.
Also had to investigate window initialization.
The levelwin (with need_redraw) is created as ROOTWIN relativly early,
but an eventloop on the ROOTWIN is dispatched in display_nhmap for the
first time. So that is the point from which on we draw the level.